// src/app/model/materiaprima.ts
export class MateriaPrima {
    producto: string;
    cantidad: number;
    precioVenProd: number;
    materiaprimapri: string;
    cantMatPri: number;
    unidad: number;
    valorMatPri: number;
    matPrima2: string;
    valorMatPri2: number;
    matPrima3: string;
    valorMatPri3: number;
    matPrima4: string;
    valorMatPri4: number;
    matPrima5: string;
    valorMatPri5: number;
    ManoObra: string;
    valorMao: number;
    cif: string;
    valorCif: number;
    precioCompra: number;
    precioVenta: number;
    participacion: number;
    porcentaje: number;
}